- renamed WizardStep.ts to wizard-step.interface.ts
