+++ src/components/components/wizard-completion-step.component.ts
@@ -5,7 +5,7 @@
 import {Component, ContentChild, EventEmitter, forwardRef, HostBinding, Inject, Input, Output} from '@angular/core';
 import {MovingDirection} from '../util/moving-direction.enum';
 import {WizardComponent} from './wizard.component';
-import {WizardStep} from '../util/WizardStep';
+import {WizardStep} from '../util/wizard-step.interface';
 import {WizardStepTitleDirective} from '../directives/wizard-step-title.directive';
 
 /**
