+++ src/components/components/wizard.component.ts
@@ -1,6 +1,6 @@
 import {AfterContentInit, Component, ContentChildren, HostBinding, Input, QueryList} from '@angular/core';
 import {MovingDirection} from '../util/moving-direction.enum';
-import {WizardStep} from '../util/WizardStep';
+import {WizardStep} from '../util/wizard-step.interface';
 import {isBoolean} from 'util';
 
 /**
