+++ src/components/directives/go-to-step.directive.ts
@@ -6,7 +6,7 @@ import {Directive, Output, HostListener, EventEmitter, Input, Optional} from '@a
 import {WizardComponent} from '../components/wizard.component';
 import {isStepOffset, StepOffset} from '../util/step-offset.interface';
 import {isNumber, isString} from 'util';
-import {isWizardStep, WizardStep} from '../util/WizardStep';
+import {isWizardStep, WizardStep} from '../util/wizard-step.interface';
 
 /**
  * The `goToStep` directive can be used to navigate to a given step.
